Refatorado mais codigo: limpeza de includes, using namespaces em lugares errados, colisões vazias removidas e outras coisas. Criado mais contantes, dano de MummyWeapon e MummyRangedWeapon é definido pelo MummyBuilder.
git-svn-id: http://xp-dev.com/svn/horus_eye/trunk@315 863e7ce6-8861-447e-8dcb-03987b876380

+++ src/game/sprites/itembuilder.h
@@ -35,7 +35,7 @@ class ItemBuilder {
     };
     class IncreaseSightEvent : public ItemEvent {
       public:
-        IncreaseSightEvent (int additional_sight) : additional_sight_(additional_sight) {}
+        IncreaseSightEvent (float additional_sight) : additional_sight_(additional_sight) {}
         bool Use (Hero *);
 
       private:
